Do You Feel Like Tim Do?
I implore you read this.

This is a first draft. It is by no means finished. There are therefore a few things worth mentioning.

Because it is not finished, I have not added to nor updated the readme. There are no Trainers, nor Reductions, as of yet. This is Expert Pro strings only.

Furthermore, there are one or two sections I've left blank on Guitar as I don't know what exactly is played.

Bass should be mostly okay. It looks fine in origin, but there might be a few pitch errors or charting mistakes, or easier/better ways of playing things.

Guitar, like I said, besides the blank spots (of which it's only 1 or 2 bars or so), check pitches. Especially chords. A lot of the chords are probably not quite right. The solos should be mostly okay, except the really fast parts might have a few mistakes occasionally. Though I do find figuring out solos generally easier than chords, so they should mostly be okay.

Then obviously if the chords need changing, root notes might be wrong and chord names might not be entirely correct.

Finally, for Guitar, I haven't used any lanes on the faster sections. If any sections should be made into lanes please let me know (and also remind me exactly how to lane stuff). I've also marked bends on Guitar (no bends on Bass that I can remember), but it's hard to tell when and where something is a bend or a very fast HOPO or something.

I think that's more or less everything. This took a long time to make, and there aren't even any reductions or trainers yet!