Added Positive Force. (#147)
* Added Positive Force.

Also changed my previous upgrades' difficulty values

* Slide and taps

Fixed an incorrect slide on bass and also changed the main riff to taps

* Fixes Made

Ok, so REALLY good feedback there, thank you. Was literally racking my brain on how to make that part of the bass chart playable.

I removed the chord names for guitar and also forced chord numbering. I felt like that was a good alternative.

I changed the tuning of the bass so that those two parts would actually be hittable, and it makes so much sense now that I am looking at it.

I did my best for the difficulties. I am very very bad at pro strings, so it's hard for me to accurately label the difficulty of it. Let me know if you think that's good.

* Return of Root Notes

if a root note is there, but you can't see it

is it really there at all?

* Gave my old rbn upgrades real difficulties