The Metal Pro Strings upgrade (#166)
* The Metal Pro Strings Upgrade

This always felt the most guitar hero-y of all the real guitar parts I have ever learned (minus that fingerstyle section), so it feels good to have it here.

* Update RB DLC README.md with tunings

* Fix slight typo on RB DLC README.md

* Adjusted difficulties

Somehow the lower Magma real_bass difficulty still appeared as the wrong tier, not really sure what happened there.

* Removed hide markers from fingerstyle section

Not touching the custom chord names because they work as-is, but I did notice in the process that C6 and Am7 were hidden when they really shouldn't be.