Six Tim One Pro strings upgrade
Everything is done, but a few things.

The Guitar stems have both guitars playing for most of the song, and they play different parts (one lead, one rhythm). So it was very difficult to make the Guitar charts.

Slides on guitar had to be reversed, as they weren't correct in Onyx. Hopefully they're fine in game. I think the issue was because Onyx didn't like the open string in some of the chords in the main riff.

I'm not happy about trainers for both guitar and bass, but I didn't know how else to do them. The song is simple, and I took the hardest repetition of each riff I could find on guitar and bass, and put the trainers there, rather than just have them at the start of verses. With the trainers, one or two might be too long, and some might not be needed or too similar to other trainers.

For such a simple song, there were issues making the upgrade, but hopefully it is mostly okay.