Added Airbrushed (#205)
* Added Airbrushed

Added Airbrushed. Tuning E Standard.

Also tweaked Positive Force's chord markers just to make SURE that they work.

* Fixed chord, altered forced numberings to exclude octaves, and raise bass difficulty

I left the forced numbering in for the octaves that are near octaves with different fingerings

* Oops forgot the left hand position changes

* Remove force numbering and reduce hard further

* Change to octaves

In addition to verse 1b (and the repeated instance later in the song) I also changed another of the same type of chord which sounded like it could've been the same sort of issue

* More octaves

* Fixed some chords again lol

Because I was rushing lol.
Anyway, I went and changed a few of the other chords too. Being on the topic of sliding, it kinda sounded like they were slides too